// src/generated/models/Cr108_messagesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr108_messagesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr108_messagesesstatecode = keyof typeof Cr108_messagesesstatecode;
export const Cr108_messagesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr108_messagesesstatuscode = keyof typeof Cr108_messagesesstatuscode;

export interface Cr108_messagesesBase {
  cr108_directed_to: string;
  cr108_message: string;
  cr108_message_key: string;
  cr108_messagesid: string;
  "cr108_request@odata.bind": string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr108_messagesesstatecode;
  statuscode?: Cr108_messagesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr108_messageses extends Cr108_messagesesBase {
  cr108_requestname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr108_request?: object;
  _cr108_request_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}